internal/parser: move ParsePattern into the parser
Addresses IDX-142
