+++ types/patttern_test.go
@@ -19,89 +19,6 @@ func TestPatternFromBuilder(t *testing.T) {
 	})
 }
 
-func TestParsePattern(t *testing.T) {
-	t.Parallel()
-	a := String("a")
-	tests := []struct {
-		input   string
-		wantOk  bool
-		want    Pattern
-		wantErr string
-	}{
-		{"", true, NewPattern(), ""},
-		{"a", true, NewPattern(a), ""},
-		{"*", true, NewPattern(Wildcard()), ""},
-		{"*a", true, NewPattern(Wildcard(), a), ""},
-		{"a*", true, NewPattern(a, Wildcard()), ""},
-		{"**", true, NewPattern(Wildcard()), ""},
-		{"**a", true, NewPattern(Wildcard(), a), ""},
-		{"a**", true, NewPattern(a, Wildcard()), ""},
-		{"*a*", true, NewPattern(Wildcard(), a, Wildcard()), ""},
-		{"**a**", true, NewPattern(Wildcard(), a, Wildcard()), ""},
-		{"abra*ca", true, NewPattern(String("abra"), Wildcard(), String("ca")), ""},
-		{"abra**ca", true, NewPattern(String("abra"), Wildcard(), String("ca")), ""},
-		{"*abra*ca", true, NewPattern(Wildcard(), String("abra"), Wildcard(), String("ca")), ""},
-		{"abra*ca*", true, NewPattern(String("abra"), Wildcard(), String("ca"), Wildcard()), ""},
-		{"*abra*ca*", true, NewPattern(Wildcard(), String("abra"), Wildcard(), String("ca"), Wildcard()), ""},
-		{"*abra*ca*dabra", true, NewPattern(Wildcard(), String("abra"), Wildcard(), String("ca"), Wildcard(), String("dabra")), ""},
-		{`*abra*c\**da\*bra`, true, NewPattern(Wildcard(), String("abra"), Wildcard(), String("c*"), Wildcard(), String("da*bra")), ""},
-		{`\u`, false, Pattern{}, "bad unicode rune"},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.input, func(t *testing.T) {
-			t.Parallel()
-			got, err := ParsePattern(tt.input)
-			if err != nil {
-				testutil.Equals(t, tt.wantOk, false)
-				testutil.Equals(t, err.Error(), tt.wantErr)
-			} else {
-				testutil.Equals(t, tt.wantOk, true)
-				testutil.Equals(t, got, tt.want)
-			}
-		})
-	}
-}
-
-func TestMatch(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		pattern string
-		target  string
-		want    bool
-	}{
-		{`""`, "", true},
-		{`""`, "hello", false},
-		{`"*"`, "hello", true},
-		{`"e"`, "hello", false},
-		{`"*e"`, "hello", false},
-		{`"*e*"`, "hello", true},
-		{`"hello"`, "hello", true},
-		{`"hello*"`, "hello", true},
-		{`"*h*llo*"`, "hello", true},
-		{`"h*e*o"`, "hello", true},
-		{`"h*e**o"`, "hello", true},
-		{`"h*z*o"`, "hello", false},
-
-		{`"\u{210d}*"`, "ℍello", true},
-		{`"\u{210d}*"`, "Hello", false},
-
-		{`"\*\**\*\*"`, "**foo**", true},
-		{`"\*\**\*\*"`, "**bar**", true},
-		{`"\*\**\*\*"`, "*bar*", false},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.pattern+":"+tt.target, func(t *testing.T) {
-			t.Parallel()
-			pat, err := ParsePattern(tt.pattern[1 : len(tt.pattern)-1])
-			testutil.OK(t, err)
-			got := pat.Match(String(tt.target))
-			testutil.Equals(t, got, tt.want)
-		})
-	}
-}
-
 func TestJSON(t *testing.T) {
 	t.Parallel()
 	tests := []struct {
